+++ src/crud_views_polymorphic/lib/__init__.py
@@ -1,18 +1,23 @@
 from crud_views_polymorphic.lib.create import PolymorphicCreateView, PolymorphicCreateViewPermissionRequired
 from crud_views_polymorphic.lib.create_select import (
+    PolymorphicContentTypeForm,
     PolymorphicCreateSelectView,
     PolymorphicCreateSelectViewPermissionRequired,
 )
+from crud_views_polymorphic.lib.delete import PolymorphicDeleteView, PolymorphicDeleteViewPermissionRequired
 from crud_views_polymorphic.lib.detail import PolymorphicDetailView, PolymorphicDetailViewPermissionRequired
 from crud_views_polymorphic.lib.update import PolymorphicUpdateView, PolymorphicUpdateViewPermissionRequired
 
 __all__ = [
-    "PolymorphicDetailView",
-    "PolymorphicDetailViewPermissionRequired",
+    "PolymorphicContentTypeForm",
     "PolymorphicCreateSelectView",
     "PolymorphicCreateSelectViewPermissionRequired",
     "PolymorphicCreateView",
     "PolymorphicCreateViewPermissionRequired",
+    "PolymorphicDeleteView",
+    "PolymorphicDeleteViewPermissionRequired",
+    "PolymorphicDetailView",
+    "PolymorphicDetailViewPermissionRequired",
     "PolymorphicUpdateView",
     "PolymorphicUpdateViewPermissionRequired",
 ]